+++ raylib/utils_windows.go
@@ -16,7 +16,7 @@ import (
 func SetTraceLog(typeFlags int) {
 	logTypeFlags = typeFlags
 
-	ctypeFlags := (C.uchar)(typeFlags)
+	ctypeFlags := (C.int)(typeFlags)
 	C.SetTraceLogLevel(ctypeFlags)
 }
